CADExchange: 修复 CADExchange 当前编译存在一批非功能性 warning，同时 feature_builder_accessor_examples_20260318.md 作为后续 CAD 映射参考还不够完整。warning 主要分两类：示例程序里的 Unicode 输出字符导致 MSVC 代码页 warning，以及 accessor 里 size_t 到 int 的隐式窄化和 topologyIndex 这类 legacy 兼容字段反复触发 deprecated warning。文档层面虽然已经迁移到 SweepExtent，但 RevolveAccessor 的完整读取接口、Builder::Extent 的推荐范围写法、以及草图为何必须保留局部变量来承接 localID 的边界说明仍不够清晰，容易让后续桥接开发继续沿用旧心智或直接改底层数据。
【问题详情与根因】
CADExchange 当前编译存在一批非功能性 warning，同时 feature_builder_accessor_examples_20260318.md 作为后续 CAD 映射参考还不够完整。warning 主要分两类：示例程序里的 Unicode 输出字符导致 MSVC 代码页 warning，以及 accessor 里 size_t 到 int 的隐式窄化和 topologyIndex 这类 legacy 兼容字段反复触发 deprecated warning。文档层面虽然已经迁移到 SweepExtent，但 RevolveAccessor 的完整读取接口、Builder::Extent 的推荐范围写法、以及草图为何必须保留局部变量来承接 localID 的边界说明仍不够清晰，容易让后续桥接开发继续沿用旧心智或直接改底层数据。

【修复方案与实现】
将 RecommendedApproach.cpp 和 PartReconstructionDemo.cpp 里的状态符号统一改成 ASCII 文本标签 [OK]/[WARN]/[ERROR]/[FAIL]，消除代码页相关 warning；为 SketchAccessor::GetSegmentCount/GetConstraintCount 和 ModelAccessor::GetFeatureCount 增加显式转换，消除 size_t 到 int 的窄化 warning。对 ReferenceAccessor、ReferenceBuilder、UnifiedSerialization、TinyXMLSerializer 这几处明确属于 legacy topologyIndex 兼容访问的文件，使用局部 MSVC warning push/pop 收口 4996，避免兼容层持续污染构建输出。文档方面，补充了 RevolveAccessor 当前完整的 extent 读取接口，增加 Builder::Extent 的 ThroughAll/UpToNext/ThroughAllBothSides 推荐示例，新增 Revolve 通过引用终止范围的链式示例，并明确说明哪些场景必须落局部变量、哪些场景应坚持单条 builder 链到 Build，同时强调不要手工修改 SweepExtent 底层字段。

+++ service/accessors/ModelAccessor.h
@@ -53,7 +53,9 @@ public:
   /**
    * @brief 获取模型中的特征数量。
    */
-  int GetFeatureCount() const { return Data()->GetFeatures().size(); }
+  int GetFeatureCount() const {
+    return static_cast<int>(Data()->GetFeatures().size());
+  }
 
   /**
    * @brief 按索引获取通用特征访问器。
